ml/embedding-pipeline/requirements.txt
sentence-transformers==2.2.2
torch==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1
pgvector==0.2.3

